#28: Restructure code.
