+++ integration_test/app_test.dart
@@ -7,11 +7,10 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:integration_test/integration_test.dart';
 
 import 'package:hackathon_flutterturkiye_org/main.dart' as app;
 
-void main() => run(_testMain);
+void main() => _testMain();
 
 void _testMain() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
