+++ lib/core/utils/assets.dart
@@ -1,0 +1,3 @@
+class ImageAssets {
+  static const String logo = 'assets/images/logo.png';
+}
