+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hackathon_flutterturkiye_org/core/styles/theme_helper.dart';
-import 'package:hackathon_flutterturkiye_org/pages/home_page.dart';
+import 'package:hackathon_flutterturkiye_org/flutter_festival_home_page.dart';
 
 void main() {
   runApp(MyApp());
@@ -12,7 +12,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Festival',
       theme: ThemeHelper.themeData,
-      home: HomePage(title: 'Flutter Festival'),
+      home: FlutterFestivalHomePage(title: 'Flutter Festival'),
       debugShowCheckedModeBanner: false,
     );
   }
