--- lib/utils/assets.dart
@@ -1,6 +1,0 @@
-class Assets {
-  static const String logo = 'assets/images/logo.png';
-  static const String footerBackground = 'assets/images/footer_bg.jpg';
-  static const String speakerPlaceHolder =
-      'assets/images/speaker_placeholder.png';
-}
